% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step1_fct_get_landuse.R
\name{download_elevation}
\alias{download_elevation}
\title{Download Elevation Raster}
\usage{
download_elevation(shape, inapp = F)
}
\arguments{
\item{shape}{An `sf` object defining the area of interest.}

\item{inapp}{Logical. If `TRUE`, shows a notification in a Shiny app upon failure.}
}
\value{
A cropped elevation raster (`SpatRaster`).
}
\description{
Downloads and crops a global elevation raster (2.5 arc-min resolution) to the extent of an sf object.
}
