% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step1_fct_get_landuse.R
\name{download_footprint}
\alias{download_footprint}
\title{Download Human Footprint Raster}
\usage{
download_footprint(shape, year = 2009, inapp = F)
}
\arguments{
\item{shape}{An `sf` object defining the area of interest.}

\item{year}{Numeric year for the footprint dataset (default = 2009).}

\item{inapp}{Logical. If `TRUE`, shows a notification in a Shiny app upon failure.}
}
\value{
A cropped footprint raster (`SpatRaster`).
}
\description{
Downloads and crops a global human footprint raster to the extent of an sf object.
}
