% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step2_utils_get_points.R
\name{get_cities}
\alias{get_cities}
\title{Get OpenStreetMap city data within a bounding box}
\usage{
get_cities(bbox, crs = 4326, in_app = FALSE)
}
\arguments{
\item{bbox}{a vector containing coordinates of the search area as c(xmin, ymin, xmax, ymax)}

\item{crs}{the four-digit code for the coordinate reference system of the data. Default is EPSG:4326}
}
\value{
a `sf` containing cities within the bounding box
#' @export get_cities
}
\description{
A function returning a shapefile of the cities within a bounding box. Uses
OpenStreetMap API (requires internet to access). Any location tagged
"city", "borough", "suburb", "quarter", "village", "town", "hamlet" is
retrieved. Cities are returned as point coordinate objects.
}
