% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step2_utils_get_points.R
\name{get_land_area}
\alias{get_land_area}
\title{Get land area within a shape}
\usage{
get_land_area(shape)
}
\arguments{
\item{shape}{a `sf` object representing the spatial extent within which to
calculate the land area.}
}
\value{
an `sf` object representing the valid land area within the shape.
}
\description{
Get land area within a shape
}
\examples{
  shape <- sf::st_as_sfc(sf::st_bbox(c(xmin = -120, xmax = -119, ymin = 35, ymax = 36), crs = 4326))
  result <- get_land_area(shape)

}
