% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step2_utils_get_points.R
\name{get_random_points}
\alias{get_random_points}
\title{Get random points within a polygon or bounding box}
\usage{
get_random_points(
  area,
  n_points,
  min_dist = 0,
  road_dist = 0,
  city_dist = 0,
  crs = 4326,
  in_app = FALSE
)
}
\arguments{
\item{area}{either a vector containing coordinates of the search area as c(xmin, ymin, xmax, ymax)
or a sf object containing a polygon representing the search area.}

\item{n_points}{int. number of points to find}

\item{min_dist}{int. minimum distance between sampled points}

\item{road_dist}{int. distance to the nearest road in meters.}

\item{city_dist}{int. points must be further from cities than this distance, in meters.}

\item{crs}{the four-digit code for the coordinate reference system of the data. Default is EPSG:4326}

\item{in_app}{if in shiny version, adds notifications}
}
\value{
a `sf` object containing lat, long coordinates for random points located
within the provided polygon and bounding box.
}
\description{
Returns a shapefile of points within a bounding box or polygon.
Only includes points located on land, and filters points located close to roads or
cities based on the distances provided.
}
\examples{
  bbox <- c(-85, 29, -82, 31)
  result <- get_random_points(area=bbox, n_points=10)

}
