% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step2_utils_get_points.R
\name{get_roads}
\alias{get_roads}
\title{Get OpenStreetMap roads data within a bounding box}
\usage{
get_roads(bbox, crs = 4326, in_app = FALSE)
}
\arguments{
\item{bbox}{a vector containing coordinates of the search area as c(xmin, ymin, xmax, ymax)}

\item{crs}{the four-digit code for the coordinate reference system of the data. Default is EPSG:4326}

\item{in_app}{if in shiny version, adds notifications}
}
\value{
a `sf` containing the highways within a bounding box
#' @export get_roads
}
\description{
A function returning a shapefile of the roads within a bounding box. Uses
OpenStreetMap API (requires internet to access). Roads returned as line objects.
Only roads tagged "highway" are retrieved.
}
