% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step2_utils_get_points.R
\name{get_village_points}
\alias{get_village_points}
\title{Get village points within a polygon  or bounding box}
\usage{
get_village_points(area, crs = 4326, in_app = FALSE)
}
\arguments{
\item{area}{either a vector containing coordinates of the search area as c(xmin, ymin, xmax, ymax)
or a sf object containing a polygon representing the search area.}

\item{crs}{the four-digit code for the coordinate reference system of the data. Default is EPSG:4326.}

\item{in_app}{if in shiny version, adds notifications}
}
\value{
a `sf` containing points of the cities in the area
}
\description{
Returns a shapefile with the latitude and longitude of cities located
within a bounding box or polygon. Uses the OpenStreetMap API to find anywhere
tagged "city", "suburb", "village", "town", "hamlet". If 'max_pop' is provided, only
cities with a population smaller than the provided value are returned.
}
\examples{
  bbox <- sf::st_bbox(c(xmin = -120, xmax = -119, ymin = 35, ymax = 36), crs = 4326)
  result <- get_village_points(area=bbox)

}
