% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step1_fct_get_landuse.R
\name{get_worldcover}
\alias{get_worldcover}
\title{Download ESA WorldCover data}
\usage{
get_worldcover(shape, tile_limit = 3, inapp = FALSE, coarse_res = 100)
}
\arguments{
\item{shape}{sf. sf object containing the area of interest}

\item{inapp}{boolean. prints messages to shiny app if in app}
}
\value{
a SpatRaster of landuse for the shape area.
}
\description{
Downloads 10 m resolution land cover data
}
