% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step3_fct_lc_analysis.R
\name{siteRasterStats}
\alias{siteRasterStats}
\title{Extract raster statistics around site buffers}
\usage{
siteRasterStats(in_df, raster, dist, progress = FALSE)
}
\arguments{
\item{in_df}{A data frame with a geometry column (sf object) and site metadata.}

\item{raster}{A SpatRaster (either categorical or continuous).}

\item{dist}{Radius (in CRS units) around each site to extract raster values.}

\item{progress}{Logical. Show progress if running inside a Shiny app.}
}
\value{
A long-format data frame with summary statistics or patch metrics.
}
\description{
Extract raster statistics around site buffers
}
