% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitetool-package.R
\docType{package}
\name{sitetool-package}
\alias{sitetool}
\alias{sitetool-package}
\title{sitetool: SiteTool}
\description{
R Shiny application for geospatially informed field site selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BioDivHealth/sitetool}
  \item Report bugs at \url{https://github.com/BioDivHealth/sitetool/issues}
}

}
\author{
\strong{Maintainer}: Natalie Imirzian \email{nimirzian@gmail.com}

}
\keyword{internal}
