% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssanalyzer-package.R
\docType{package}
\name{sitetool-package}
\alias{sitetool}
\alias{sitetool-package}
\title{sitetool: SITE_SELECTION_ANALYZER}
\description{
Shiny application to analyze landcover for site selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://https://github.com/BioDivHealth/ss-analyzer}
  \item Report bugs at \url{https://https://github.com/BioDivHealth/ss-analyzer/issues}
}

}
\author{
\strong{Maintainer}: Natalie Imirzian \email{nimirzian@gmail.com}

}
\keyword{internal}
